% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callcnv.R
\name{callcnv}
\alias{callcnv}
\title{Call copy number variants}
\usage{
callcnv(controldir, experimentdir, outputdir = getwd(), species = NULL)
}
\arguments{
\item{controldir}{The parent directory of the sample directories.}

\item{experimentdir}{The parent directory of sample on which to investigate copy numbers.}

\item{outputdir}{The directory in which to place copy number call.}

\item{species}{A character string indicating the ID of the species to use
(e.g. "1" for hg19,
"2" for hg38...). If NULL, an interactive menu will prompt the user.}
}
\value{
A data frame containing copy number calls.
}
\description{
This function calls copy number variants from sample BAM files compared to
control BAM files. It assumes that BAM files are stored in separate folders
as is created by fastqProcession(). This function requires that control BAM
files are provided. Once complete, it creates a CSV file containing copy
number information.
}
