% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfcallSNP.R
\name{callsnp}
\alias{callsnp}
\title{Call single nucleotide polymorphism variants}
\usage{
callsnp(bam, threads = 4L, outputdir = getwd())
}
\arguments{
\item{bam}{A list of paths to BAM files}

\item{threads}{The number of cores to use.}

\item{outputdir}{The output directory for the VCF file.}
}
\value{
A list of file paths to the VCF files.
}
\description{
This function calls SNP variants from BAM files. The results are formatted
into a VCF file and the ID column is populated with dbSNP IDs.
}
