% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Count genes}
\usage{
count(
  dir = getwd(),
  groups = NULL,
  outputdir = getwd(),
  threads = 4L,
  paired = FALSE,
  species = NULL
)
}
\arguments{
\item{dir}{The parent directory of the sample groups.}

\item{groups}{Folder names of the sample groups. The default is all folders in dir.}

\item{outputdir}{Output directory of CSV file. By default, it uses the current
working directory.}

\item{threads}{Number of cores to use.}

\item{paired}{Indicates whether the samples are from paired-end reads.}

\item{species}{A character string indicating the ID of the species to use
(e.g. "1" for hg19,
"2" for hg38...). If NULL, an interactive menu will prompt the user.}
}
\value{
A data frame containing annotated gene counts.
}
\description{
This function counts reads of gene regions between sample groups.
It assumes that sample BAM files are ordered in a directory structure such as
"group/sample/" as processfastq() would order it. It outputs a CSV file
showing gene counts. Works similarly to geneExpression(), but outputs count
data instead of differential expression data.
}
