% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_Counting.R
\name{counts}
\alias{counts}
\title{Count genes}
\usage{
counts(
  dir = getwd(),
  groups = NULL,
  outputdir = getwd(),
  threads = 4L,
  paired = FALSE
)
}
\arguments{
\item{dir}{The parent directory of the sample groups.}

\item{groups}{Folder names of the sample groups. The default is all folders in dir.}

\item{outputdir}{Output directory of CSV file.}

\item{threads}{Number of cores to use.}

\item{paired}{Indicates whether the samples are from paired-end reads.}
}
\value{
A data frame containing gene counts.
}
\description{
This function counts reads of gene regions between sample groups.
It assumes that sample BAM files are ordered in a directory structure such as
"group/sample/" as processfastq() would order it. It outputs a CSV file
showing gene counts. Works similarly to expression(), but outputs count
data instead of differential expression data.
}
