% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_supported_genomes}
\alias{list_supported_genomes}
\title{List supported genomes in exvar}
\usage{
list_supported_genomes()
}
\value{
A data.frame with columns: ID, Package and Directory
}
\description{
This function returns a data frame with available species IDs and corresponding
genome packages
}
