% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processfastq.R
\name{processfastq}
\alias{processfastq}
\title{Convert a FASTQ file to an indexed BAM}
\usage{
processfastq(
  file = list_files_with_exts(dir = dir, exts = "fastq"),
  dir = getwd(),
  paired = FALSE,
  threads = 4L,
  molecule = "RNA",
  species = NULL
)
}
\arguments{
\item{file}{A list of paths to FASTQ files. If no paths are entered, defaults
to all fastq files in dir.}

\item{dir}{Output directory. By default, it uses the current working directory.}

\item{paired}{A logical indicating whether the data is single or pair-end.}

\item{threads}{The number of cores to use in the process.}

\item{molecule}{A character string indicating either DNA or RNA samples.}

\item{species}{A character string indicating the ID of the species to use
(e.g. "1" for hg19,
"2" for hg38...). If NULL, an interactive menu will prompt the user.}
}
\value{
A list of \code{BamFile} objects.
}
\description{
This function takes in FASTQ files and performs quality control before
aligning to a reference genome. It assumes paired-end samples are of the same
file name with an underscore (_) and a number to signify different reads of
the same sample. Each sample's outputs will be stored in a separate
directory.
}
