% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirement.R
\name{requirement}
\alias{requirement}
\title{requirement}
\usage{
requirement(all = FALSE)
}
\arguments{
\item{all}{Logical. If \code{FALSE} (default), prompts the user to select a species from a predefined list.
If \code{TRUE}, installs dependencies for all available species (1 to 8) without prompting.}
}
\value{
The function does not return a value but installs necessary packages.
}
\description{
Install required packages based on operating system and selected species
}
\examples{
exvar::requirement()
}
