% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizcnv.R
\name{vizcnv}
\alias{vizcnv}
\title{vizcnv}
\usage{
vizcnv(cnvdata, species)
}
\arguments{
\item{cnvdata}{Path to the CNVs data file}

\item{species}{Species name from the following list: "Homo Sapiens", "Mus Musculus","Arabidopsis Thaliana" ,
"Drosophila Melanogaster", "Danio rerio", "Rattus norvegicus", or "Saccharomyces Cerevisiae" .}
}
\description{
vizcnv
}
