% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizexp.R
\name{vizexp}
\alias{vizexp}
\title{vizexp}
\usage{
vizexp(genecount, metadata, species)
}
\arguments{
\item{genecount}{Path to the gene count file}

\item{metadata}{Path to the  metadata file}

\item{species}{Species name ( "Homo Sapiens", "Mus Musculus","Arabidopsis Thaliana" ,
"Drosophila Melanogaster", "Danio rerio", "Rattus norvegicus", or "Saccharomyces Cerevisiae" )}
}
\description{
vizexp
}
