% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callcnv.R
\name{callcnv}
\alias{callcnv}
\title{Call copy number variants}
\usage{
callcnv(controldir, experimentdir, outputdir = getwd())
}
\arguments{
\item{controldir}{The parent directory of the sample directories.}

\item{experimentdir}{The parent directory of sample on which to investigate copy numbers.}

\item{outputdir}{The directory in which to place copy number call.}
}
\value{
A data frame containing copy number calls.
}
\description{
This function calls copy number variants from sample BAM files compared to
control BAM files. It assumes that BAM files are stored in separate folders
as is created by processfastq(). This function requires that control BAM
files are provided. Once complete, it creates a CSV file containing copy
number information.
}
