% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCNV.R
\name{callcnv}
\alias{callcnv}
\title{Call copy number variants}
\usage{
callcnv(
  controldir,
  control = NULL,
  experimentdir,
  experiment = NULL,
  bed,
  outputdir = getwd()
)
}
\arguments{
\item{controldir}{The parent directory of the sample directories.}

\item{control}{The names of the folders in which control BAM files are. If NULL, all folders in controldir will be checked for BAM files.}

\item{experimentdir}{The parent directory of sample on which to investigate copy numbers.}

\item{experiment}{The names of the folders in which sample BAM files are. If NULL, all folders in experimentdir will be checked for BAM files.}

\item{bed}{A character string indicating BED file path or a TxDb object from which to extract a BED file.}

\item{outputdir}{The directory in which to place copy number call.}
}
\value{
A data frame containing copy number calls.
}
\description{
This function calls copy number variants from sample BAM files compared to
control BAM files. It assumes that BAM files are stored in separate folders
as is created by fastqProcession(). This function requires that control BAM
files are provided. Once complete, it creates a CSV file containing copy
number information.
}
