% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{geneExpression}
\alias{geneExpression}
\title{Analyse differential gene expression}
\usage{
geneExpression(
  dir = getwd(),
  groups = NULL,
  outputdir = getwd(),
  threads = 4L,
  paired = FALSE
)
}
\arguments{
\item{dir}{The parent directory of the sample groups.}

\item{groups}{Folder names of the sample groups. The default is all folders in dir.}

\item{outputdir}{Output directory of CSV file.}

\item{threads}{Number of cores to use.}

\item{paired}{Indicates whether the samples are from paired-end reads.}
}
\value{
A data frame list containing all of the differential expression comparison.
}
\description{
This function analyses differentially expressed genes between sample groups.
It assumes that sample BAM files are ordered in a directory structure such as
"group/sample/" as fastqProcession() would order it. There should be more
than one sample per group or else differential expression analysis won't work
It outputs a CSV file showing differential expression (ordered by p-value).
It works similarly to gene_Counting(), but then further analyses those counts
to obtain differential expression data.
}
