% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizexp.R
\name{vizexp}
\alias{vizexp}
\title{vizexp
This function is a shiny app to visualize gene expression data.}
\usage{
vizexp(genecount, metadata)
}
\arguments{
\item{genecount}{Path to the gene count file}

\item{metadata}{Path to the  metadata file}
}
\description{
vizexp
This function is a shiny app to visualize gene expression data.
}
